Added the tests for zero code change
