+++ tests/zero_code_change/test_pytorch_integration.py
@@ -16,7 +16,6 @@ import torch
 import torch.nn as nn
 import torch.nn.functional as F
 import torch.optim as optim
-from packaging import version
 from tests.zero_code_change.pt_utils import Net, get_dataloaders
 
 # First Party
@@ -92,48 +91,6 @@ def test_pytorch(script_mode, use_loss_module):
                 for name in hook.collection_manager.get("losses").tensor_names
             ]
         )
-
-
-@pytest.fixture()
-def pytorch_framework_override(monkeypatch):
-    import smdebug.pytorch.utils
-
-    monkeypatch.setattr(smdebug.pytorch.utils, "PT_VERSION", version.parse("1.14"))
-    return
-
-
-def test_pytorch_with_unsupported_version(pytorch_framework_override, use_loss_module=False):
-    smd.del_hook()
-
-    sim_class = SagemakerSimulator
-    with sim_class() as sim:
-        trainloader, testloader = get_dataloaders()
-        net = Net()
-        criterion = nn.CrossEntropyLoss()
-        optimizer = optim.SGD(net.parameters(), lr=0.001, momentum=0.9)
-
-        for i, data in enumerate(trainloader, 0):
-            # get the inputs; data is a list of [inputs, labels]
-            inputs, labels = data
-
-            # zero the parameter gradients
-            optimizer.zero_grad()
-
-            # forward + backward + optimize
-            outputs = net(inputs)
-            if use_loss_module:
-                loss = criterion(outputs, labels)
-            else:
-                loss = F.cross_entropy(outputs, labels)
-            loss.backward()
-            optimizer.step()
-            if i == 499:  # print every 2000 mini-batches
-                break
-
-        print("Finished Training")
-
-        hook = smd.get_hook()
-        assert hook == None
 
 
 if __name__ == "__main__":
